iOS 17 state restoration crash fix
