+++ Zotero/Scenes/Master/MasterContainerViewController.swift
@@ -40,54 +40,24 @@ final class MasterContainerViewController: UINavigationController {
     private static let minVisibleBottomHeight: CGFloat = 200
     private let disposeBag: DisposeBag
 
-    private var bottomContainer: UIView!
+    lazy var bottomController: DraggableViewController? = {
+        return coordinatorDelegate?.createBottomController()
+    }()
+    private var bottomContainer: UIView?
     private var bottomYConstraint: NSLayoutConstraint!
     private var bottomContainerBottomConstraint: NSLayoutConstraint!
     // Current position of bottom container
     private var bottomPosition: BottomPosition
     // Previous position of bottom container. Used to return to previous position when drag handle is tapped.
     private var previousBottomPosition: BottomPosition?
-    private var didAppear: Bool
     // Used to calculate position and velocity when dragging
     private var initialBottomMinY: CGFloat?
     private var keyboardHeight: CGFloat = 0
-    
-    private var _bottomController: DraggableViewController?
-    var bottomController: DraggableViewController? {
-        if _bottomController != nil {
-            return _bottomController
-        }
-        guard let bottomController = coordinatorDelegate?.bottomController() else { return nil }
-        setupBottomController()
-        _bottomController = bottomController
-        return _bottomController
-        
-        func setupBottomController() {
-            bottomController.view.translatesAutoresizingMaskIntoConstraints = false
-            // Since the instance keeps a strong reference to the bottomController, its view is simply added as a subview.
-            // Adding bottomController as a child view controller, would mess up the navigation stack.
-            bottomContainer.addSubview(bottomController.view)
-            let bottomControllerHeight = bottomController.view.heightAnchor.constraint(greaterThanOrEqualToConstant: 100)
-            bottomControllerHeight.priority = .required
-            let bottomControllerBottom = bottomController.view.bottomAnchor.constraint(equalTo: bottomContainer.bottomAnchor)
-            bottomControllerBottom.priority = UILayoutPriority(999)
-            NSLayoutConstraint.activate([
-                // bottom controller view
-                bottomController.view.topAnchor.constraint(equalTo: bottomContainer.topAnchor, constant: Self.bottomControllerHandleHeight),
-                bottomController.view.leadingAnchor.constraint(equalTo: bottomContainer.leadingAnchor),
-                bottomController.view.trailingAnchor.constraint(equalTo: bottomContainer.trailingAnchor),
-                bottomControllerHeight,
-                bottomControllerBottom
-            ])
-        }
-    }
 
-    private weak var coordinatorDelegate: MasterContainerCoordinatorDelegate?
+    weak var coordinatorDelegate: MasterContainerCoordinatorDelegate?
     
-    init(coordinatorDelegate: MasterContainerCoordinatorDelegate) {
-        self.coordinatorDelegate = coordinatorDelegate
+    init() {
         self.bottomPosition = .default
-        self.didAppear = false
         self.disposeBag = DisposeBag()
         super.init(nibName: nil, bundle: nil)
     }
@@ -103,9 +73,11 @@ final class MasterContainerViewController: UINavigationController {
         view.backgroundColor = .clear
         setupView()
         setupKeyboardObserving()
-        showBottomSheet(false)
+        setBottomSheet(hidden: true)
         
         func setupView() {
+            guard let bottomController = bottomController else { return }
+
             let bottomPanRecognizer = UIPanGestureRecognizer()
             bottomPanRecognizer.delegate = self
             bottomPanRecognizer.rx.event
@@ -147,6 +119,15 @@ final class MasterContainerViewController: UINavigationController {
             separator.backgroundColor = .opaqueSeparator
             bottomContainer.addSubview(separator)
 
+            bottomController.view.translatesAutoresizingMaskIntoConstraints = false
+            // Since the instance keeps a strong reference to the bottomController, its view is simply added as a subview.
+            // Adding bottomController as a child view controller, would mess up the navigation stack.
+            bottomContainer.addSubview(bottomController.view)
+
+            let bottomControllerHeight = bottomController.view.heightAnchor.constraint(greaterThanOrEqualToConstant: 100)
+            bottomControllerHeight.priority = .required
+            let bottomControllerBottom = bottomController.view.bottomAnchor.constraint(equalTo: bottomContainer.bottomAnchor)
+            bottomControllerBottom.priority = UILayoutPriority(999)
             let bottomYConstraint = bottomContainer.topAnchor.constraint(equalTo: view.topAnchor)
             let bottomContainerBottomConstraint = view.bottomAnchor.constraint(equalTo: bottomContainer.bottomAnchor)
 
@@ -170,17 +151,23 @@ final class MasterContainerViewController: UINavigationController {
                 separator.heightAnchor.constraint(equalToConstant: 1 / UIScreen.main.scale),
                 separator.topAnchor.constraint(equalTo: bottomContainer.topAnchor),
                 separator.leadingAnchor.constraint(equalTo: bottomContainer.leadingAnchor),
-                separator.trailingAnchor.constraint(equalTo: bottomContainer.trailingAnchor)
+                separator.trailingAnchor.constraint(equalTo: bottomContainer.trailingAnchor),
+                // bottom controller view
+                bottomController.view.topAnchor.constraint(equalTo: bottomContainer.topAnchor, constant: Self.bottomControllerHandleHeight),
+                bottomController.view.leadingAnchor.constraint(equalTo: bottomContainer.leadingAnchor),
+                bottomController.view.trailingAnchor.constraint(equalTo: bottomContainer.trailingAnchor),
+                bottomControllerHeight,
+                bottomControllerBottom
             ])
 
             self.bottomYConstraint = bottomYConstraint
             self.bottomContainerBottomConstraint = bottomContainerBottomConstraint
-            
+
             func toolbarDidPan(recognizer: UIPanGestureRecognizer) {
                 switch recognizer.state {
                 case .began:
-                    initialBottomMinY = self.bottomContainer.frame.minY
-                    bottomController?.disablePanning()
+                    initialBottomMinY = self.bottomContainer?.frame.minY ?? 0
+                    bottomController.disablePanning()
 
                 case .changed:
                     guard let initialBottomMinY else { return }
@@ -218,7 +205,7 @@ final class MasterContainerViewController: UINavigationController {
                     }
 
                     initialBottomMinY = nil
-                    bottomController?.enablePanning()
+                    bottomController.enablePanning()
 
                 case .cancelled, .possible:
                     break
@@ -310,22 +297,15 @@ final class MasterContainerViewController: UINavigationController {
         }
     }
 
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        guard !didAppear else { return }
-        didAppear = true
+    override func viewIsAppearing(_ animated: Bool) {
+        super.viewIsAppearing(animated)
+
+        updateBottomPosition()
         if let splitViewController {
             // Split view controller collapsed status when the app launches is correct here, so it's used to show/hide bottom sheet for the first appearance.
             // The app may be launched in collapsed mode, if it was in such mode the last time it was moved to background.
-            showBottomSheet(!splitViewController.isCollapsed)
+            setBottomSheet(hidden: splitViewController.isCollapsed)
         }
-    }
-
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-
-        guard !didAppear else { return }
-        updateBottomPosition()
     }
 
     override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
@@ -340,7 +320,7 @@ final class MasterContainerViewController: UINavigationController {
     }
         
     override func collapseSecondaryViewController(_ secondaryViewController: UIViewController, for splitViewController: UISplitViewController) {
-        showBottomSheet(false)
+        setBottomSheet(hidden: true)
         super.collapseSecondaryViewController(secondaryViewController, for: splitViewController)
         // The search bar is hidden when the app goes to background for unknown reason. This is a workaround to reset it if needed when
         // the app returns to active state.
@@ -350,7 +330,7 @@ final class MasterContainerViewController: UINavigationController {
     }
 
     override func separateSecondaryViewController(for splitViewController: UISplitViewController) -> UIViewController? {
-        showBottomSheet(true)
+        setBottomSheet(hidden: false)
         guard topViewController?.isKind(of: UINavigationController.self) == true else {
             // When separating from an initially collapsed split view controller, the detail view controller is not yet set.
             coordinatorDelegate?.showDefaultCollection()
@@ -360,14 +340,11 @@ final class MasterContainerViewController: UINavigationController {
     }
 
     // MARK: - Bottom panning
-    private func showBottomSheet(_ show: Bool) {
-        if show, bottomController != nil {
-            bottomContainer.isHidden = false
-            additionalSafeAreaInsets = UIEdgeInsets(top: 0, left: 0, bottom: 16, right: 0)
-        } else {
-            bottomContainer.isHidden = true
-            additionalSafeAreaInsets = .zero
-        }
+
+    private func setBottomSheet(hidden: Bool) {
+        guard let container = bottomContainer else { return }
+        container.isHidden = hidden
+        additionalSafeAreaInsets = hidden ? .zero : UIEdgeInsets(top: 0, left: 0, bottom: 16, right: 0)
     }
 
     private func set(bottomPosition: BottomPosition, containerHeight: CGFloat) {
